refactor(plugins): move `Identifier` from audio_bridge into common.rs
