+++ jarust_plugins/src/audio_bridge/common.rs
@@ -1,17 +1,6 @@
 use serde::Deserialize;
-use serde::Serialize;
 
-/// Rooms and Participants Identifier.
-///
-/// Identifier should be by default unsigned integer, unless configured otherwise in the audiobridge config.
-#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum Identifier {
-    /// String Identifier
-    String(String),
-    /// Unsigned Integer Identifier
-    Uint(u64),
-}
+use crate::common::Identifier;
 
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, Deserialize)]
 pub struct Participant {
