+++ jarust_plugins/src/lib.rs
@@ -1,3 +1,6 @@
+pub use common::AttachPluginParams;
+pub use common::Identifier;
+
 #[cfg(feature = "echo_test")]
 pub mod echo_test;
 
@@ -8,5 +11,3 @@ pub mod audio_bridge;
 pub mod video_room;
 
 pub mod common;
-
-pub use common::AttachPluginParams;
